+++ demisto_sdk/commands/validate/validate_manager.py
@@ -1,3 +1,4 @@
+import json
 import os
 import re
 from configparser import ConfigParser, MissingSectionHeaderError
@@ -7,7 +8,7 @@ import click
 from demisto_sdk.commands.common import tools
 from demisto_sdk.commands.common.configuration import Configuration
 from demisto_sdk.commands.common.constants import (
-    CONTENT_ENTITIES_DIRS, KNOWN_FILE_STATUSES, PACKS_DIR,
+    API_MODULES_PACK, CONTENT_ENTITIES_DIRS, KNOWN_FILE_STATUSES, PACKS_DIR,
     PACKS_INTEGRATION_NON_SPLIT_YML_REGEX, PACKS_PACK_IGNORE_FILE_NAME,
     PACKS_PACK_META_FILE_NAME, PACKS_SCRIPT_NON_SPLIT_YML_REGEX,
     TESTS_DIRECTORIES, FileType)
@@ -50,7 +51,8 @@ from demisto_sdk.commands.common.hook_validations.structure import \
     StructureValidator
 from demisto_sdk.commands.common.hook_validations.widget import WidgetValidator
 from demisto_sdk.commands.common.tools import (filter_packagify_changes,
-                                               find_type,
+                                               find_type, get_api_module_ids,
+                                               get_api_module_integrations_set,
                                                get_content_release_identifier,
                                                get_pack_name,
                                                get_pack_names_from_files,
@@ -64,7 +66,7 @@ class ValidateManager:
     def __init__(self, is_backward_check=True, prev_ver=None, use_git=False, only_committed_files=False,
                  print_ignored_files=False, skip_conf_json=True, validate_id_set=False, file_path=None,
                  validate_all=False, is_external_repo=False, skip_pack_rn_validation=False, print_ignored_errors=False,
-                 silence_init_prints=False, no_docker_checks=False, skip_dependencies=False):
+                 silence_init_prints=False, no_docker_checks=False, skip_dependencies=False, id_set_path=None):
         # General configuration
         self.skip_docker_checks = False
         self.no_configuration_prints = silence_init_prints
@@ -84,6 +86,9 @@ class ValidateManager:
         # Class constants
         self.handle_error = BaseValidator(print_as_warnings=print_ignored_errors).handle_error
         self.file_path = file_path
+        if not id_set_path:
+            id_set_path = 'Tests/id_set.json'
+        self.id_set_path = id_set_path
         self.branch_name = ''
         self.changes_in_schema = False
         self.check_only_schema = False
@@ -615,8 +620,8 @@ class ValidateManager:
 
         changed_packs = modified_packs.union(added_packs).union(changed_meta_packs)
 
-        if not os.path.isfile('Tests/id_set.json'):
-            IDSetCreator(print_logs=False, output='Tests/id_set.json').create_id_set()
+        if not os.path.isfile(self.id_set_path):
+            IDSetCreator(print_logs=False, output=self.id_set_path).create_id_set()
 
         for pack in changed_packs:
             raise_version = False
@@ -625,7 +630,7 @@ class ValidateManager:
                 raise_version = True
             valid_pack_files.add(self.validate_pack_unique_files(
                 pack_path, self.get_error_ignore_list(pack), should_version_raise=raise_version,
-                id_set_path='Tests/id_set.json'))
+                id_set_path=self.id_set_path))
 
         return all(valid_pack_files)
 
@@ -693,6 +698,15 @@ class ValidateManager:
                                                                                    FileType.TEST_PLAYBOOK,
                                                                                    FileType.TEST_SCRIPT,
                                                                                    FileType.DOC_IMAGE})
+        if API_MODULES_PACK in packs_that_should_have_new_rn:
+            if not os.path.isfile(self.id_set_path):
+                IDSetCreator(print_logs=False, output=self.id_set_path).create_id_set()
+            with open(self.id_set_path, 'r') as conf_file:
+                id_set = json.load(conf_file)
+            api_module_set = get_api_module_ids(changed_files)
+            integrations = get_api_module_integrations_set(api_module_set, id_set.get('integrations', []))
+            packs_that_should_have_new_rn = packs_that_should_have_new_rn.union(
+                set(map(lambda integration: integration.get('pack'), integrations)))
 
         # new packs should not have RN
         packs_that_should_have_new_rn = packs_that_should_have_new_rn - self.new_packs